+++ packages/sst/src/constructs/SsrSite.ts
@@ -49,6 +49,10 @@ import {
   CacheQueryStringBehavior,
   CacheHeaderBehavior,
   CacheCookieBehavior,
+  OriginRequestPolicy,
+  Function as CfFunction,
+  FunctionCode as CfFunctionCode,
+  FunctionEventType as CfFunctionEventType,
 } from "aws-cdk-lib/aws-cloudfront";
 import { ICertificate } from "aws-cdk-lib/aws-certificatemanager";
 import { AwsCliLayer } from "aws-cdk-lib/lambda-layer-awscli";
@@ -59,6 +63,7 @@ import { App } from "./App.js";
 import { Stack } from "./Stack.js";
 import { Logger } from "../logger.js";
 import { SSTConstruct, isCDKConstruct } from "./Construct.js";
+import { NodeJSProps } from "./Function.js";
 import { EdgeFunction } from "./EdgeFunction.js";
 import {
   BaseSiteDomainProps,
@@ -77,7 +82,6 @@ import {
 } from "./util/functionBinding.js";
 import { SiteEnv } from "../site-env.js";
 import { useProject } from "../project.js";
-import { VisibleError } from "../error.js";
 
 const __dirname = url.fileURLToPath(new URL(".", import.meta.url));
 
@@ -87,6 +91,7 @@ export type SsrBuildConfig = {
   clientBuildVersionedSubDir: string;
 };
 
+export interface SsrSiteNodeJSProps extends NodeJSProps {}
 export interface SsrDomainProps extends BaseSiteDomainProps {}
 export interface SsrSiteReplaceProps extends BaseSiteReplaceProps {}
 export interface SsrCdkDistributionProps extends BaseSiteCdkDistributionProps {}
@@ -170,6 +175,10 @@ export interface SsrSiteProps {
    * ```
    */
   runtime?: "nodejs14.x" | "nodejs16.x" | "nodejs18.x";
+  /**
+   * Used to configure nodejs function properties
+   */
+  nodejs?: SsrSiteNodeJSProps;
   /**
    * Attaches the given list of permissions to the SSR function. Configuring this property is equivalent to calling `attachPermissions()` after the site is created.
    * @example
@@ -255,10 +264,15 @@ export interface SsrSiteProps {
  */
 export class SsrSite extends Construct implements SSTConstruct {
   public readonly id: string;
-  protected props: Omit<SsrSiteProps, "path"> & {
-    path: string;
-    timeout: number | Duration;
-    memorySize: number | Size;
+  protected props: SsrSiteProps & {
+    path: Exclude<SsrSiteProps["path"], undefined>;
+    runtime: Exclude<SsrSiteProps["runtime"], undefined>;
+    timeout: Exclude<SsrSiteProps["timeout"], undefined>;
+    memorySize: Exclude<SsrSiteProps["memorySize"], undefined>;
+    waitForInvalidation: Exclude<
+      SsrSiteProps["waitForInvalidation"],
+      undefined
+    >;
   };
   private doNotDeploy: boolean;
   protected buildConfig: SsrBuildConfig;
@@ -703,15 +717,7 @@ export class SsrSite extends Construct implements SSTConstruct {
   }
 
   private createFunctionPermissionsForRegional() {
-    const { permissions } = this.props;
-
     this.bucket.grantReadWrite(this.serverLambdaForRegional!.role!);
-    if (permissions) {
-      attachPermissionsToRole(
-        this.serverLambdaForRegional!.role as Role,
-        permissions
-      );
-    }
   }
 
   private createFunctionPermissionsForEdge() {
@@ -808,12 +814,14 @@ export class SsrSite extends Construct implements SSTConstruct {
 
     return {
       viewerProtocolPolicy: ViewerProtocolPolicy.REDIRECT_TO_HTTPS,
+      functionAssociations: this.buildBehaviorFunctionAssociations(),
       origin: new HttpOrigin(Fn.parseDomainName(fnUrl.url)),
       allowedMethods: AllowedMethods.ALLOW_ALL,
       cachedMethods: CachedMethods.CACHE_GET_HEAD_OPTIONS,
       compress: true,
       cachePolicy:
         cdk?.serverCachePolicy ?? this.createCloudFrontServerCachePolicy(),
+      originRequestPolicy: this.createCloudFrontServerOriginRequestPolicy(),
       ...(cfDistributionProps.defaultBehavior || {}),
     };
   }
@@ -826,12 +834,14 @@ export class SsrSite extends Construct implements SSTConstruct {
 
     return {
       viewerProtocolPolicy: ViewerProtocolPolicy.REDIRECT_TO_HTTPS,
+      functionAssociations: this.buildBehaviorFunctionAssociations(),
       origin,
       allowedMethods: AllowedMethods.ALLOW_ALL,
       cachedMethods: CachedMethods.CACHE_GET_HEAD_OPTIONS,
       compress: true,
       cachePolicy:
         cdk?.serverCachePolicy ?? this.createCloudFrontServerCachePolicy(),
+      originRequestPolicy: this.createCloudFrontServerOriginRequestPolicy(),
       ...(cfDistributionProps.defaultBehavior || {}),
       // concatenate edgeLambdas
       edgeLambdas: [
@@ -843,6 +853,22 @@ export class SsrSite extends Construct implements SSTConstruct {
         ...(cfDistributionProps.defaultBehavior?.edgeLambdas || []),
       ],
     };
+  }
+
+  private buildBehaviorFunctionAssociations() {
+    return [
+      {
+        eventType: CfFunctionEventType.VIEWER_REQUEST,
+        function: new CfFunction(this, "CloudFrontFunction", {
+          code: CfFunctionCode.fromInline(`
+function handler(event) {
+  var request = event.request;
+  request.headers["x-forwarded-host"] = request.headers.host;
+  return request;
+}`),
+        }),
+      },
+    ];
   }
 
   protected buildDistributionStaticFileBehaviors(
@@ -877,7 +903,7 @@ export class SsrSite extends Construct implements SSTConstruct {
     return staticsBehaviours;
   }
 
-  protected createCloudFrontServerCachePolicy(): CachePolicy {
+  protected createCloudFrontServerCachePolicy() {
     return new CachePolicy(this, "ServerCache", {
       queryStringBehavior: CacheQueryStringBehavior.all(),
       headerBehavior: CacheHeaderBehavior.none(),
@@ -889,6 +915,15 @@ export class SsrSite extends Construct implements SSTConstruct {
       enableAcceptEncodingGzip: true,
       comment: "SST server response cache policy",
     });
+  }
+
+  protected createCloudFrontServerOriginRequestPolicy() {
+    // CloudFront's Managed-AllViewerExceptHostHeader policy
+    return OriginRequestPolicy.fromOriginRequestPolicyId(
+      this,
+      "ServerOriginRequestPolicy",
+      "b689b0a8-53d0-40ab-baf2-68738e2966ac"
+    );
   }
 
   private createCloudFrontInvalidation() {