+++ packages/sst/src/runtime/handlers/node.ts
@@ -23,10 +23,10 @@ export const useNodeHandler = Context.memo(async () => {
     shouldBuild: (input) => {
       const result = cache[input.functionID];
       if (!result) return false;
-      const relative = path.relative(
-        project.paths.root,
-        input.file
-      ).split(path.sep).join(path.posix.sep);
+      const relative = path
+        .relative(project.paths.root, input.file)
+        .split(path.sep)
+        .join(path.posix.sep);
       return Boolean(result.metafile?.inputs[relative]);
     },
     canHandle: (input) => input.startsWith("nodejs"),
@@ -117,11 +117,16 @@ export const useNodeHandler = Context.memo(async () => {
       }
 
       const { external, ...override } = nodejs.esbuild || {};
+      const forceExternal = [
+        "sharp",
+        "pg-native",
+        ...(isESM || input.props.runtime === "nodejs18.x" ? [] : ["aws-sdk"]),
+      ];
       const options: BuildOptions = {
         entryPoints: [file],
         platform: "node",
         external: [
-          ...(isESM || input.props.runtime === "nodejs18.x" ? [] : ["aws-sdk"]),
+          ...forceExternal,
           ...(nodejs.install || []),
           ...(external || []),
         ],
@@ -134,7 +139,7 @@ export const useNodeHandler = Context.memo(async () => {
           ? {
               format: "esm",
               target: "esnext",
-              mainFields: isESM ? ["module", "main"] : undefined,
+              mainFields: ["module", "main"],
               banner: {
                 js: [
                   `import { createRequire as topLevelCreateRequire } from 'module';`,
@@ -164,56 +169,88 @@ export const useNodeHandler = Context.memo(async () => {
         const result = await esbuild.build(options);
 
         // Install node_modules
-        if (options.external?.length) {
-          async function find(dir: string, target: string): Promise<string> {
-            if (dir === "/")
-              throw new VisibleError("Could not find a package.json file");
-            if (
-              await fs
-                .access(path.join(dir, target))
-                .then(() => true)
-                .catch(() => false)
-            )
-              return dir;
-            return find(path.join(dir, ".."), target);
-          }
+        const installPackages = [
+          ...(nodejs.install || []),
+          ...forceExternal
+            .filter((pkg) => pkg !== "aws-sdk")
+            .filter((pkg) => !external?.includes(pkg))
+            .filter((pkg) =>
+              Object.values(result.metafile?.inputs || {}).some(({ imports }) =>
+                imports.some(({ path }) => path === pkg)
+              )
+            ),
+        ];
 
-          if (input.mode === "deploy" && nodejs.install) {
-            const src = await find(parsed.dir, "package.json");
-            const json = JSON.parse(
-              await fs
-                .readFile(path.join(src, "package.json"))
-                .then((x) => x.toString())
-            );
-            fs.writeFile(
-              path.join(input.out, "package.json"),
-              JSON.stringify({
-                dependencies: Object.fromEntries(
-                  nodejs.install?.map((x) => [x, json.dependencies?.[x] || "*"])
-                ),
+        // TODO bubble up the warnings
+        const warnings: string[] = [];
+        Object.entries(result.metafile?.inputs || {}).forEach(
+          ([inputPath, { imports }]) =>
+            imports
+              .filter(({ path }) => path.includes("sst/constructs"))
+              .forEach(({ path }) => {
+                warnings.push(
+                  `You are importing from "${path}" in "${inputPath}". Did you mean to import from "sst/node"?`
+                );
               })
-            );
-            await new Promise<void>((resolve) => {
-              const process = exec("npm install", {
-                cwd: input.out,
-              });
-              process.on("exit", () => resolve());
-            });
-          }
+        );
 
-          if (input.mode === "start") {
-            const dir = path.join(
-              await find(parsed.dir, "package.json"),
-              "node_modules"
+        async function find(dir: string, target: string): Promise<string> {
+          if (dir === "/")
+            throw new VisibleError("Could not find a package.json file");
+          if (
+            await fs
+              .access(path.join(dir, target))
+              .then(() => true)
+              .catch(() => false)
+          )
+            return dir;
+          return find(path.join(dir, ".."), target);
+        }
+
+        if (input.mode === "deploy" && installPackages) {
+          const src = await find(parsed.dir, "package.json");
+          const json = JSON.parse(
+            await fs
+              .readFile(path.join(src, "package.json"))
+              .then((x) => x.toString())
+          );
+          fs.writeFile(
+            path.join(input.out, "package.json"),
+            JSON.stringify({
+              dependencies: Object.fromEntries(
+                installPackages.map((x) => [x, json.dependencies?.[x] || "*"])
+              ),
+            })
+          );
+          const cmd = ["npm install"];
+          if (installPackages.includes("sharp")) {
+            cmd.push(
+              "--platform=linux",
+              input.props.architecture === "arm_64"
+                ? "--arch=arm64"
+                : "--arch=x64"
             );
-            try {
-              await fs.symlink(
-                path.resolve(dir),
-                path.resolve(path.join(input.out, "node_modules")),
-                "dir"
-              );
-            } catch {}
           }
+          await new Promise<void>((resolve) => {
+            const process = exec(cmd.join(" "), {
+              cwd: input.out,
+            });
+            process.on("exit", () => resolve());
+          });
+        }
+
+        if (input.mode === "start") {
+          const dir = path.join(
+            await find(parsed.dir, "package.json"),
+            "node_modules"
+          );
+          try {
+            await fs.symlink(
+              path.resolve(dir),
+              path.resolve(path.join(input.out, "node_modules")),
+              "dir"
+            );
+          } catch {}
         }
 
         cache[input.functionID] = result;
@@ -223,13 +260,20 @@ export const useNodeHandler = Context.memo(async () => {
         };
       } catch (ex: any) {
         const result = ex as BuildResult;
+        if ("errors" in result) {
+          return {
+            type: "error",
+            errors: result.errors.flatMap((x) => [
+              Colors.bold(x.text),
+              x.location?.file || "",
+              Colors.dim(x.location?.line, "│", x.location?.lineText),
+            ]),
+          };
+        }
+
         return {
           type: "error",
-          errors: result.errors.flatMap((x) => [
-            Colors.bold(x.text),
-            x.location?.file || "",
-            Colors.dim(x.location?.line, "│", x.location?.lineText),
-          ]),
+          errors: [ex.toString()],
         };
       }
     },