+++ packages/astro-sst/CHANGELOG.md
@@ -1,5 +1,35 @@
 # astro-sst
 
+## 2.1.21
+
+## 2.1.20
+
+## 2.1.19
+
+## 2.1.18
+
+## 2.1.17
+
+## 2.1.16
+
+## 2.1.15
+
+## 2.1.14
+
+## 2.1.13
+
+## 2.1.12
+
+## 2.1.11
+
+## 2.1.10
+
+### Patch Changes
+
+- [`d2e7d097d`](https://github.com/serverless-stack/sst/commit/d2e7d097d186c68c5ebde1bdd1c56a9de4f31787) Thanks [@fwang](https://github.com/fwang)! - SsrSite: set "x-forwarded-host" using CloudFront function
+
+## 2.1.9
+
 ## 2.1.8
 
 ## 2.1.7